+++ src/components/Slider/sliderChangelogData.ts
@@ -16,6 +16,7 @@ export const changelogData: ChangelogData[] = [
     affects: ["Styles"],
     notes: [
       "Changes theme name from `CustomSlider` to `ReservoirSlider` for consistency.",
+      "Removes `isDarkMode` in favor of Chakra's `_dark` conditional key.",
     ],
   },
   {